backend/requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
pyodbc
langchain
openai
python-dotenv
